+++ src/components/common/SelectDropdown.vue
@@ -179,7 +179,6 @@ watch(selectValue, (newValue: string) => {
     }
 
     .select-chevron-icon {
-      color: var(--kui-color-text-neutral-strong, $kui-color-text-neutral-strong) !important;
       flex-shrink: 0;
       height: var(--kui-icon-size-30, $kui-icon-size-30) !important;
       width: var(--kui-icon-size-30, $kui-icon-size-30) !important;