+++ src/utils/strings.ts
@@ -6,3 +6,19 @@
 export function removeTrailingSlash(str: string) {
   return str.replace(/\/$/, '')
 }
+
+export const stringifyUnknownValue = (candidate: unknown) => {
+  let stringToRender = ''
+  if (Array.isArray(candidate)) {
+    // can't do a simple Array.toString because items of type Object are
+    // converted to [Object Object], hence we need to stringify object items
+    candidate.forEach((example, index) => {
+      stringToRender +=
+        typeof(example) === 'object' ? JSON.stringify(example) : example +
+        `${index < candidate.length - 1 ? ', ' : ''}`
+    })
+  } else if (typeof candidate === 'object') {
+    stringToRender = JSON.stringify(candidate)
+  }
+  return stringToRender || candidate
+}